Update twoAgents.py
In this project, there are two agents placed in an environment. They will find an optimal route (placed in list) to a particular location, 
pick up an item then take it to a designated drop off location (placed in a list) while following an optimal route within the 
environment.
If both agents are already in the pick up location, both are randomly selected to take the item to the drop off location while taking the 
optimal route.
If either of the agents is already in the pick up location, that agent will take the item to the drop off location while taking the 
optimal route.
Any agent closer to the pick up location will perform the duty while taking the optimal routes. 
If both agents are placed in the same location (not the pick up location) or if their respective reward values to pickup location are equal, 
both are randomly selected to perform the duty.
See TwoAgent.PNG in Images repository for diagram description of the environment.