Update optimal_route1.py
In this RL project, the agent randomly chooses a targeted location while we choose a starting location in an input prompt for the agent and the 
agent returns an optimal route to the targeted location in a list. optimal route is return if the locations (both starting and targeted) are 
situated in the environment. 
See  optimal_route_image1.png in Image repository  file that contains the diagram description of the environment.